+++ lighthouse-core/lib/traces/tracing-processor.js
@@ -11,168 +11,8 @@ const BASE_RESPONSE_LATENCY = 16;
 const SCHEDULABLE_TASK_TITLE = 'TaskQueueManager::ProcessTaskFromWorkQueue';
 const SCHEDULABLE_TASK_TITLE_ALT = 'ThreadControllerImpl::DoWork';
 const LHError = require('../errors');
-const {taskGroups, taskNameToGroup} = require('../task-groups');
-
-/** @typedef {import('../task-groups.js').TaskGroup} TaskGroup */
 
 class TraceProcessor {
-  /**
-   * @param {LH.TraceEvent} event
-   * @param {TaskNode} [parent]
-   * @return {TaskNode}
-   */
-  static _createNewTaskNode(event, parent) {
-    const newTask = {
-      event,
-      startTime: event.ts,
-      endTime: event.ph === 'X' ? event.ts + Number(event.dur || 0) : NaN,
-      parent: parent,
-      children: [],
-
-      // These properties will be filled in later
-      group: taskGroups.Other,
-      attributableURL: undefined,
-      duration: NaN,
-      selfTime: NaN,
-    };
-
-    if (parent) {
-      parent.children.push(newTask);
-    }
-
-    return newTask;
-  }
-
-  /**
-   * @param {LH.TraceEvent[]} traceEvents
-   * @return {TaskNode[]}
-   */
-  static _createTasksFromEvents(traceEvents) {
-    const {startedInPageEvt} = TraceProcessor.findTracingStartedEvt(traceEvents);
-
-    /** @type {TaskNode[]} */
-    const tasks = [];
-    /** @type {TaskNode|undefined} */
-    let currentTask;
-
-    for (const event of traceEvents) {
-      // Only look at main thread events
-      if (event.pid !== startedInPageEvt.pid || event.tid !== startedInPageEvt.tid) continue;
-      // Only look at X (Complete), B (Begin), and E (End) events as they have most data
-      if (event.ph !== 'X' && event.ph !== 'B' && event.ph !== 'E') continue;
-
-      // Update currentTask based on the elapsed time.
-      // The next event may be after currentTask has ended.
-      while (
-        currentTask &&
-        Number.isFinite(currentTask.endTime) &&
-        currentTask.endTime <= event.ts
-      ) {
-        currentTask = currentTask.parent;
-      }
-
-      // If we don't have a current task, start a new one.
-      if (!currentTask) {
-        // We can't start a task with an end event
-        if (event.ph === 'E') {
-          throw new Error('Fatal trace logic error');
-        }
-
-        currentTask = TraceProcessor._createNewTaskNode(event);
-        tasks.push(currentTask);
-
-        continue;
-      }
-
-      if (event.ph === 'X' || event.ph === 'B') {
-        // We're starting a nested event, create it as a child and make it the currentTask
-        const newTask = TraceProcessor._createNewTaskNode(event, currentTask);
-        tasks.push(newTask);
-        currentTask = newTask;
-      } else {
-        if (currentTask.event.ph !== 'B') {
-          throw new Error('Fatal trace logic error');
-        }
-
-        // We're ending an event, update the end time and the currentTask to its parent
-        currentTask.endTime = event.ts;
-        currentTask = currentTask.parent;
-      }
-    }
-
-    return tasks;
-  }
-
-  /**
-   * @param {TaskNode} task
-   * @return {number}
-   */
-  static _computeRecursiveSelfTime(task) {
-    const childTime = task.children
-      .map(TraceProcessor._computeRecursiveSelfTime)
-      .reduce((sum, child) => sum + child, 0);
-    task.duration = task.endTime - task.startTime;
-    task.selfTime = task.duration - childTime;
-    return task.duration;
-  }
-
-  /**
-   * @param {TaskNode} task
-   * @param {string} [parentURL]
-   */
-  static _computeRecursiveAttributableURL(task, parentURL) {
-    const argsData = task.event.args.data || {};
-    const stackFrames = argsData.stackTrace || [{url: undefined}];
-    const taskURL = argsData.url || (stackFrames[0] && stackFrames[0].url);
-
-    task.attributableURL = parentURL || taskURL;
-    task.children.forEach(child =>
-      TraceProcessor._computeRecursiveAttributableURL(child, task.attributableURL));
-  }
-
-  /**
-   * @param {TaskNode} task
-   * @param {TaskGroup} [parentGroup]
-   */
-  static _computeRecursiveTaskGroup(task, parentGroup) {
-    const group = taskNameToGroup[task.event.name];
-    task.group = group || parentGroup || taskGroups.Other;
-    task.children.forEach(child => TraceProcessor._computeRecursiveTaskGroup(child, task.group));
-  }
-
-  /**
-   *
-   * @param {LH.TraceEvent[]} traceEvents
-   * @return {TaskNode[]}
-   */
-  static getMainThreadTasks(traceEvents) {
-    const tasks = TraceProcessor._createTasksFromEvents(traceEvents);
-
-    // Compute the recursive properties we couldn't compute earlier, starting at the toplevel tasks
-    for (const task of tasks) {
-      if (task.parent) continue;
-
-      TraceProcessor._computeRecursiveSelfTime(task);
-      TraceProcessor._computeRecursiveAttributableURL(task);
-      TraceProcessor._computeRecursiveTaskGroup(task);
-    }
-
-    const firstTs = (tasks[0] || {startTime: 0}).startTime;
-    for (const task of tasks) {
-      task.startTime = (task.startTime - firstTs) / 1000;
-      task.endTime = (task.endTime - firstTs) / 1000;
-      task.duration /= 1000;
-      task.selfTime /= 1000;
-
-      // sanity check that we have selfTime which captures all other timing data
-      if (!Number.isFinite(task.selfTime)) {
-        throw new Error('Invalid task timing data');
-      }
-    }
-
-    return tasks;
-  }
-
   /**
    * Calculate duration at specified percentiles for given population of
    * durations.
@@ -391,19 +231,6 @@ class TraceProcessor {
  * @prop {number} start
  * @prop {number} end
  * @prop {number} duration
- */
-
-/**
- * @typedef TaskNode
- * @prop {LH.TraceEvent} event
- * @prop {TaskNode[]} children
- * @prop {TaskNode|undefined} parent
- * @prop {number} startTime
- * @prop {number} endTime
- * @prop {number} duration
- * @prop {number} selfTime
- * @prop {string|undefined} attributableURL
- * @prop {TaskGroup} group
  */
 
 module.exports = TraceProcessor;
